Refactor cacheing of invites
The old system was whenever an invite button or host button was
clicked, and ajax request was sent to the server to add or remove
that user to the appropriate list.  Now, the user makes all the
changes in the modal form for inviting people, then when the form is
closed, one big chunk of data on the status of each user is sent via
and ajax request to the cache.