Basic Dockerization
App runs in a docker container. Added a dockerfile and a heroku.yml
to allow the image to be build on Heroku.  The app still runs using
the django dev server though, that's the next step, getting gunicorn
set up.

Also added a bit of security by moving some things over to environment
variables to be set either locally or on heroku for dev vs prod
builds