Fix Selenium Tests
I didn't actually have to change the site, I just encountered some
weird bug with the selenium test where it won't let you log in more
than once in a test.  When I was trying to run multiple tests in
the same Test Class (using the same login method) it would throw an
error for whichever was called second.